requisito corrigido
