+++ src/App.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import './App.css';
-import SearchField from './components/SearchField';
+import mainPage from './components/mainPage';
 import Cart from './components/Cart';
 import NotFound from './components/NotFound';
 
@@ -20,7 +20,7 @@ export default class App extends Component {
       <div className="">
         <BrowserRouter>
           <Switch>
-            <Route exact path="/" component={ SearchField } />
+            <Route exact path="/" component={ mainPage } />
             <Route
               exact
               path="/cart"
