+++ src/Components/mainPage.js
@@ -2,10 +2,10 @@ import React from 'react';
 
 import { Link } from 'react-router-dom';
 import ListaCat from './ListaCat';
-import ListItems from '../componentes/ListItems';
+import ListItems from './ListItems';
 import { getProductsFromCategoryAndQuery } from '../services/api';
 
-class SearchField extends React.Component {
+class mainPage extends React.Component {
   constructor() {
     super();
 
@@ -22,19 +22,32 @@ class SearchField extends React.Component {
     });
   }
 
+  searchItems = async (param) => {
+    const searchList = await getProductsFromCategoryAndQuery('', param);
+    this.setState({
+      productsList: searchList.results,
+    });
+  }
+
   render() {
     const { productsList } = this.state;
     return (
       <div className="search">
-        <ListItems productsList={ productsList } />
+        <ListItems
+          productsList={ productsList }
+          searchItems={ this.searchItems }
+        />
+
         <p data-testid="home-initial-message">
           Digite algum termo de pesquisa ou escolha uma categoria.
         </p>
+
         <Link to="/cart" data-testid="shopping-cart-button">Cart</Link>
+
         <ListaCat getId={ this.getId } />
       </div>
     );
   }
 }
 
-export default SearchField;
+export default mainPage;
